Downloads/DeepLearningFinal3 - Final copy/requirements.txt
blinker==1.9.0
click==8.1.8
coloredlogs==15.0.1
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.12.23
humanfriendly==10.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
numpy==2.2.1
onnxruntime==1.20.1
packaging==24.2
pillow==11.0.0
protobuf==5.29.2
sympy==1.13.3
Werkzeug==3.1.3
